ingestions/requirements.txt
tweepy==4.10.0
kafka-python==2.0.2
pytz==2022.1
pymongo==4.1.1
google-api-python-client